requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
docker==3.6.0
docker-pycreds==0.4.0
idna==2.8
more-itertools==4.3.0
pathlib2==2.3.3
pkg-resources==0.0.0
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytest==4.0.2
python-ldap==3.1.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.54.0
